move type definitions
