+++ dotscripts/weekly/beeminder/frontend/src/global.d.ts
