+++ dotscripts/weekly/beeminder/frontend/src/main.ts
@@ -1,8 +1,8 @@
-import App from './App.svelte';
+import App from "./App.svelte";
 
 const app = new App({
-	target: document.body,
-	props: { }
+  target: document.body,
+  props: {},
 });
 
 export default app;
